+++ subsys/bluetooth/host/classic/conn_br_internal.h
@@ -1,0 +1,12 @@
+/** @file
+ *  @brief Internal APIs for Bluetooth classic connection handling.
+ */
+
+/*
+ * Copyright (c) 2015 Intel Corporation
+ * Copyright (c) 2021 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+int bt_hci_connect_br_cancel(struct bt_conn *conn);
